Added a hard anchor point that will allow the sliding panel to stop at the hard anchor point no matter what. Currently the user can slide past the anchor point but it will slide to the hard anchor point when the view is released.
To set the hard anchor point in the MainActivity the user must call
"layout.setHardAnchorPoint(true)".
This will override any anchor point set with "layout.setAnchorPoint()"
The hard point is calculated on line 523 - 524

The sliding panel can be dragged pass the hard anchor point but when the
view is released, the sliding panel will smooth slide back to the hard
anchor point. I do not think that this can be stopped since technically
the panel is sliding to 0. I may be wrong.

I have also added the ability to turn off the cover color that is drawn
on top of the main panel when the sliding panel is expanded. This was
something that I was asked to do so I left it in.

I also added back some of Android margin and padding calculations that
were missing from SlidingPanelLayout.class