The hard anchor point no matter what.
To set the hard anchor point in the MainActivity the user must call
"layout.setHardAnchorPoint(true)".
This will override any anchor point set with "layout.setAnchorPoint()"
The hard point is calculated on line 523 - 524

I have also added the ability to turn off the cover color that is drawn
on top of the main panel when the sliding panel is expanded. This was
something that I was asked to do so I left it in.

I also added back some of Android margin and padding calculations that
were missing from SlidingPanelLayout.class