Line 239 typo, layout_gravity to gravity
Trying to implement this library earlier I noticed i was receiving an error "layout_gravity must be set to either top or bottom" and when I set the layout_gravity to bottom noticed that I was still receiving the error and the app was still crashing out on initialization. When I went to the repo I noticed that the android:gravity was set to bottom not the android:layout_gravity. So I changed that and everything worked. So  I have changed the thrown IllegalArgumentException to say gravity instead of layout_gravity.