Added collapsePanel(boolean) and expandPanel(boolean)  methods
I am often in a situation where I need to collapse/expand the sliding
panel even though sliding is disabled so addind a collapsePanel(boolean
forced) and expandPanel(boolean forced) options helps with dealing with
this cases.

Besides from that, the original API is untouched i.e. collapsePanel()
and expandPanel() work in the exact way as before .

USAGE:

To expand the panel programmatically simply call collapsePanel(boolean
forced). If forced is false, the behaviour is identicall to the previous
collapsePanel(). If forced is true the panel will be collapsed even is
sliding is not enabled. The same is true for expanding the panel.