Negotiate with the child view, to determine if parent SlidingUpPanelLayout could take the onTouchEvent.
Use case: Assume a sliding-up-menu child is also a Scrollable (e.g. ListView/RecyclerView/ScrollView). A common requirement is
if the child has scroll to topmost, then any scroll-down should be handled by SlidingUpPanelLayout, i.e. when the
menu is on its topmost position, any scroll-down gesture would close the menu. To achieve this, The parent SlidingUpPanelLayout
needs to consult if the child has been on its topmost position. If yes, the parent would **INTERCEPT** the whole gesture.
Otherwise the parent would simply have the child to handle it.