Handle HorizontalScrollView within scrollable view
When a HorizontalScrollView is embedded within a ScrollView which is set as the panel's scrollable view, the HorizontalScrollView does not scroll properly. Since the HorizontalScrollView scrolls properly if the ScrollView is not set as the panel's scrollable view, adding a check to ignore the portion of the scrollable view that contains a HorizontalScrollView works to allow both the HorizontalScrollView and the ScrollView to function properly.